+++ components/sql/src/test/scala/pl/touk/nussknacker/sql/service/DatabaseQueryEnricherValidationTest.scala
@@ -5,6 +5,7 @@ import pl.touk.nussknacker.engine.api.TemplateRenderedPart.RenderedLiteral
 import pl.touk.nussknacker.engine.api.context.{OutputVar, ValidationContext}
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.CustomNodeError
 import pl.touk.nussknacker.engine.api.context.transformation.{DefinedEagerParameter, OutputVariableNameValue}
+import pl.touk.nussknacker.engine.api.definition.CompileTimePositiveDurationValidator
 import pl.touk.nussknacker.engine.api.typed.typing.{Typed, Unknown}
 import pl.touk.nussknacker.sql.db.query.{ResultSetStrategy, SingleResultStrategy}
 import pl.touk.nussknacker.sql.db.schema.MetaDataProviderFactory
@@ -74,6 +75,12 @@ class DatabaseQueryEnricherValidationTest extends BaseHsqlQueryEnricherTest {
       case service.FinalResults(expectedOutputContext, _, _) => expectedOutputContext.contains("out") shouldBe true
       case _                                                 => fail("Enricher does not return final results")
     }
+  }
+
+  test("should declare compile-time positive duration validation for the Cache TTL parameter") {
+    DatabaseQueryEnricher.cacheTTLParamDeclaration
+      .createParameter()
+      .validators should contain(CompileTimePositiveDurationValidator)
   }
 
   private def eagerValueParameter(value: Any) = DefinedEagerParameter(value, null)