+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/server/TapirHttpServiceFactory.scala
@@ -9,10 +9,9 @@ import pl.touk.nussknacker.ui.migrations.{MigrationApiAdapterService, MigrationS
 import pl.touk.nussknacker.ui.notifications.NotificationServiceImpl
 import pl.touk.nussknacker.ui.process.ScenarioAttachmentService
 import pl.touk.nussknacker.ui.process.label.ScenarioLabelsService
-import pl.touk.nussknacker.ui.process.livedata.LiveDataRepository
 import pl.touk.nussknacker.ui.process.newactivity.ActivityService
 import pl.touk.nussknacker.ui.process.newdeployment.DeploymentService
-import pl.touk.nussknacker.ui.process.repository.{DBIOActionRunner, ScenarioMetadataRepository}
+import pl.touk.nussknacker.ui.process.repository.ScenarioMetadataRepository
 import pl.touk.nussknacker.ui.process.version.{ScenarioGraphVersionRepository, ScenarioGraphVersionService}
 import pl.touk.nussknacker.ui.security.api.{AuthManager, NussknackerInternalUser}
 import pl.touk.nussknacker.ui.statistics.{
@@ -42,6 +41,7 @@ object TapirHttpServiceFactory {
       categories = processingTypeServicesProvider.mapValues(_.category),
       processService = processService,
       shouldExposeConfig = designerConfig.enableConfigEndpoint,
+      leadership = domainServices.leadership
     )
 
     val migrationApiHttpService = {