+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/definition/component/parameter/ParameterData.scala
@@ -7,7 +7,7 @@ import java.lang.reflect.Parameter
 import scala.reflect.ClassTag
 
 //we extract needed data from java.lang.reflect.Parameter to be able to use it e.g. for fragment parameters
-case class ParameterData(typing: TypingResult, annotations: List[_ <: Annotation]) {
+final case class ParameterData(typing: TypingResult, annotations: List[_ <: Annotation], isLazyParameter: Boolean) {
 
   def getAnnotation[T <: Annotation: ClassTag]: Option[T] = annotations.collectFirst { case e: T =>
     e
@@ -17,6 +17,6 @@ case class ParameterData(typing: TypingResult, annotations: List[_ <: Annotation
 
 object ParameterData {
 
-  def apply(parameter: Parameter, typing: TypingResult): ParameterData =
-    ParameterData(typing, parameter.getAnnotations.toList)
+  def apply(parameter: Parameter, typing: TypingResult, isLazyParameter: Boolean): ParameterData =
+    ParameterData(typing, parameter.getAnnotations.toList, isLazyParameter)
 }